feat: Update start.sh to generate rclone configuration for all unique remotes
This commit updates the start.sh script to generate rclone configuration for all unique remotes found in the environment variables that start with RCLONE_CONFIG_. It iterates over the remotes and generates the configuration file /root/.config/rclone/rclone.conf accordingly. This change improves the flexibility and scalability of the script by allowing the configuration of multiple remotes.