chore: Refactor Makefile to dynamically calculate next release version
This commit refactors the Makefile to dynamically calculate the next release version based on the current tags. It introduces a new logic that increments the patch version by 1 and appends the current date in the format "_%Y-%m-%d" to the version number. This change improves the release process by automating the versioning and ensures unique release tags.